moved the sort function into the util file
